cleanup and abstraction to new util folders and pruning of dead/unusable code...not finished
