+++ src/components/pages/Cases/CaseTable.js
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import axiosWithAuth from '../../../utils/axiosWithAuth';
 import { Link } from 'react-router-dom';
-import Plot from 'react-plotly.js';
 import Highlighter from 'react-highlight-words';
 import OrangeLine from '../../../styles/orange-line.svg';
 import {
@@ -18,6 +17,16 @@ import { Table, Space, Button, Input, Tabs, notification, Tag } from 'antd';
 import './CaseTable.less';
 import CaseDetails from '../CaseOverview/CaseDetails';
 
+// Case column utils import
+import { case_columns } from '../../../utils/case_utils/case_columns';
+
+import DecisionRateChart from './DecisionRateChart';
+
+const initialDetails = {
+  date: '5/26/2021',
+  origin_city: 'Detroit',
+};
+
 export default function CaseTable(props) {
   const [state, setState] = useState({
     searchText: '',
@@ -25,12 +34,43 @@ export default function CaseTable(props) {
     selectedRowID: [],
   });
 
-  const initialDetails = {
-    date: '5/26/2021',
-    origin_city: 'Detroit',
-  };
   const [isDetailsVisible, setIsDetailsVisible] = useState(false);
   const [detailsData, setDetailsData] = useState(initialDetails);
+
+  // state to keep track of filters being applied to the table (Initial cases section)
+  const [initialFilters, setInitialFilters] = useState([]);
+  // keeping track of filters applied to the appelate section
+  const [appFilters, setAppFilters] = useState([]);
+
+  // state to keep track of the current table being displayed
+  const [currentKey, setCurrentKey] = useState(1);
+
+  const [tempHook, setTempHook] = useState({
+    key: '',
+    value: '',
+  });
+
+  const [removing, setRemoving] = useState(false);
+
+  const [searching, setSearching] = useState(false);
+
+  const [queryValues] = useState({
+    number: '',
+    date: '',
+    judge: '',
+    case_origin_city: '',
+    case_origin_state: '',
+    filed_in_one_year: '',
+    application_type: '',
+    protected_grounds: '',
+    outcome: '',
+    country_of_origin: '',
+    gender: '',
+    type_of_violence: '',
+    indigenous_group: '',
+    applicant_language: '',
+    credible: '',
+  });
 
   const popUpDetails = rowData => {
     setDetailsData(rowData);
@@ -79,11 +119,6 @@ export default function CaseTable(props) {
   }
   //! end of my code
 
-  const [tempHook, setTempHook] = useState({
-    key: '',
-    value: '',
-  });
-  const [removing, setRemoving] = useState(false);
   const tempFunction = async (setKeys, newValue) => {
     await setKeys([newValue]);
     setRemoving(false);
@@ -195,25 +230,6 @@ export default function CaseTable(props) {
     setState({ searchText: '' });
   };
 
-  function formatDate(text) {
-    var year = text.slice(0, 4);
-    var month = text.slice(5, 7);
-    var day = text.slice(8, 10);
-    return month + '/' + day + '/' + year;
-  }
-
-  // state to keep track of filters being applied to the table (Initial cases section)
-  const [initialFilters, setInitialFilters] = useState([]);
-  // keeping track of filters applied to the appelate section
-  const [appFilters, setAppFilters] = useState([]);
-
-  // state to keep track of the current table being displayed
-  const [currentKey, setCurrentKey] = useState(1);
-
-  useEffect(() => {
-    // current use is to keep the filter state in sync.
-  }, [initialFilters]);
-
   // returns processed array of filters
   const processFilters = filters => {
     //! console.log('FILTER', filters)
@@ -268,119 +284,6 @@ export default function CaseTable(props) {
     selectedRowID,
     onChange: onSelectChange,
   };
-
-  const columns = [
-    {
-      title: 'Case Details',
-      dataIndex: 'case_row',
-      key: 'x',
-      render: text => <FileTextOutlined onClick={() => popUpDetails(text)} />,
-    },
-    {
-      title: 'Case Date',
-      dataIndex: 'date',
-      key: 'date',
-      sorter: (a, b) => a.date.localeCompare(b.date),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('date', tempHook.key),
-      render: text => formatDate(text),
-    },
-    {
-      title: 'Judge Name',
-      dataIndex: 'judge_name',
-      key: 'judge_name',
-      sorter: (a, b) => a.last_name.localeCompare(b.last_name),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('judge_name', tempHook.key),
-      render: (text, record) => (
-        <Link to={`/judge/${record.judge_id}`}>{text}</Link>
-      ),
-    },
-    {
-      title: 'Origin City',
-      dataIndex: 'case_origin_city',
-      key: 'case_origin_city',
-      sorter: (a, b) => a.case_origin_city.localeCompare(b.case_origin_city),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('case_origin_city', tempHook.key),
-    },
-    {
-      title: 'Origin State',
-      dataIndex: 'case_origin_state',
-      key: 'case_origin_state',
-      sorter: (a, b) => a.case_origin_state.localeCompare(b.case_origin_state),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('case_origin_state', tempHook.key),
-    },
-    {
-      title: 'Filed within Year',
-      dataIndex: 'filed_within_year',
-      key: 'filed_within_year',
-      sorter: (a, b) => a.filed_within_year.localeCompare(b.filed_within_year),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('filed_within_year', tempHook.key),
-    },
-    {
-      title: 'Protected Grounds',
-      dataIndex: 'protected_grounds',
-      key: 'protected_grounds',
-      sorter: (a, b) => a.protected_grounds.localeCompare(b.protected_grounds),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('protected_grounds', tempHook.key),
-    },
-    {
-      title: 'Outcome',
-      dataIndex: 'outcome',
-      key: 'outcome',
-      sorter: (a, b) => a.outcome.localeCompare(b.outcome),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('outcome', tempHook.key),
-    },
-    {
-      title: 'Country of Origin',
-      dataIndex: 'country_of_origin',
-      key: 'country_of_origin',
-      sorter: (a, b) => a.country_of_origin.localeCompare(b.country_of_origin),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('country_of_origin', tempHook.key),
-    },
-    {
-      title: 'Applicant Gender',
-      dataIndex: 'gender',
-      key: 'gender',
-      sorter: (a, b) => a.gender.localeCompare(b.gender),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('gender', tempHook.key),
-    },
-    {
-      title: 'Violence Experienced',
-      dataIndex: 'type_of_violence',
-      key: 'type_of_violence',
-      sorter: (a, b) => a.type_of_violence.localeCompare(b.type_of_violence),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('type_of_violence', tempHook.key),
-    },
-    {
-      title: 'Applicant Language',
-      dataIndex: 'applicant_language',
-      key: 'applicant_language',
-      sorter: (a, b) =>
-        a.applicant_language.localeCompare(b.applicant_language),
-      sortDirections: ['descend', 'ascend'],
-      ...getColumnSearchProps('applicant_language', tempHook.key),
-    },
-    {
-      title: 'Download PDF',
-      dataIndex: 'url',
-      key: 'url',
-      render: text => (
-        <a href={text}>
-          {' '}
-          <FilePdfOutlined />
-        </a>
-      ),
-    },
-  ];
 
   const findRowByID = (rowID, rowData) => {
     for (let i = 0; i < rowData.length; i++) {
@@ -438,26 +341,6 @@ export default function CaseTable(props) {
     }
   };
 
-  const [searching] = useState(false);
-
-  const [queryValues] = useState({
-    number: '',
-    date: '',
-    judge: '',
-    case_origin_city: '',
-    case_origin_state: '',
-    filed_in_one_year: '',
-    application_type: '',
-    protected_grounds: '',
-    outcome: '',
-    country_of_origin: '',
-    gender: '',
-    type_of_violence: '',
-    indigenous_group: '',
-    applicant_language: '',
-    credible: '',
-  });
-
   const filter = data => {
     const searchedKeys = Object.entries(queryValues).filter(
       ([k, v]) => v !== ''
@@ -479,106 +362,51 @@ export default function CaseTable(props) {
     return filteredData;
   };
 
+  //? Right now we will only return "caseData", but never "filter(caseData)"
+  //? This functionality needs to be implemented!!!!!!
   const data = searching ? filter(caseData) : caseData;
 
-  const DecisionRateChart = () => {
-    let denied = 0;
-    let granted = 0;
-    let remanded = 0;
-    let sustained = 0;
-    let terminated = 0;
+  //? To be detemined whether we should keep it here,
+  //? or delete this piece of code
+  // const CaseDataChart = () => {
+  //   let denied = 0;
+  //   let granted = 0;
+  //   let remanded = 0;
+  //   let sustained = 0;
+  //   let terminated = 0;
 
-    data.map(eachCase => {
-      if (eachCase.outcome === 'Denied') {
-        denied += 1;
-      }
-      if (eachCase.outcome === 'Granted') {
-        granted += 1;
-      }
-      if (eachCase.outcome === 'Remanded') {
-        remanded += 1;
-      }
-      if (eachCase.outcome === 'Sustained') {
-        sustained += 1;
-      }
-      if (eachCase.outcome === 'Terminated') {
-        terminated += 1;
-      }
-      return null;
-    });
+  //   data.map(eachCase => {
+  //     if (eachCase.outcome === 'Denied') {
+  //       denied += 1;
+  //     }
+  //     if (eachCase.outcome === 'Granted') {
+  //       granted += 1;
+  //     }
+  //     if (eachCase.outcome === 'Remanded') {
+  //       remanded += 1;
+  //     }
+  //     if (eachCase.outcome === 'Sustained') {
+  //       sustained += 1;
+  //     }
+  //     if (eachCase.outcome === 'Terminated') {
+  //       terminated += 1;
+  //     }
+  //     return null;
+  //   });
 
-    return (
-      <Plot
-        data={[
-          {
-            type: 'bar',
-            x: ['Granted', 'Denied', 'Remanded', 'Sustained', 'Terminated'],
-            y: [granted, denied, remanded, sustained, terminated],
-          },
-        ]}
-        useResizeHandler={true}
-        config={{
-          modeBarButtonsToRemove: [
-            'toImage',
-            'zoom2d',
-            'pan2d',
-            'select2d',
-            'lasso2d',
-            'drawclosedpath',
-            'drawopenpath',
-            'zoomIn2d',
-            'zoomOut2d',
-            'autoScale2d',
-            'hoverClosestCartesian',
-            'hoverCompareCartesian',
-            'toggleSpikelines',
-          ],
-          displaylogo: false,
-        }}
-        layout={{ width: 500, height: 300, title: 'Decision Rate' }}
-      />
-    );
-  };
-
-  const CaseDataChart = () => {
-    let denied = 0;
-    let granted = 0;
-    let remanded = 0;
-    let sustained = 0;
-    let terminated = 0;
-
-    data.map(eachCase => {
-      if (eachCase.outcome === 'Denied') {
-        denied += 1;
-      }
-      if (eachCase.outcome === 'Granted') {
-        granted += 1;
-      }
-      if (eachCase.outcome === 'Remanded') {
-        remanded += 1;
-      }
-      if (eachCase.outcome === 'Sustained') {
-        sustained += 1;
-      }
-      if (eachCase.outcome === 'Terminated') {
-        terminated += 1;
-      }
-      return null;
-    });
-
-    return (
-      <Plot
-        data={[
-          {
-            type: 'bar',
-            x: ['Granted', 'Denied', 'Remanded', 'Sustained', 'Terminated'],
-            y: [granted, denied, remanded, sustained, terminated],
-          },
-        ]}
-        layout={{ width: 500, height: 300, title: 'Case Data' }}
-      />
-    );
-  };
+  //   return (
+  //     <Plot
+  //       data={[
+  //         {
+  //           type: 'bar',
+  //           x: ['Granted', 'Denied', 'Remanded', 'Sustained', 'Terminated'],
+  //           y: [granted, denied, remanded, sustained, terminated],
+  //         },
+  //       ]}
+  //       layout={{ width: 500, height: 300, title: 'Case Data' }}
+  //     />
+  //   );
+  // };
 
   const nonAppCases = casesData.filter(item => item.appellate === false);
   // const appCases = casesData.filter(item => item.appellate === true);
@@ -631,7 +459,7 @@ export default function CaseTable(props) {
         isDetailsVisible={isDetailsVisible}
       />
       <div className="viz-container">
-        <DecisionRateChart className="casesViz" />
+        <DecisionRateChart data={data} className="casesViz" />
       </div>
 
       <div className="case-table-container">
@@ -673,40 +501,19 @@ export default function CaseTable(props) {
                 className="cases_table iCases"
                 rowSelection={rowSelection}
                 rowKey={record => record.case_id}
-                columns={columns}
+                columns={case_columns(
+                  getColumnSearchProps,
+                  tempHook,
+                  Link,
+                  popUpDetails,
+                  FileTextOutlined,
+                  FilePdfOutlined
+                )}
                 dataSource={nonAppCases}
                 onChange={changeSorter}
               />
             </div>
           </TabPane>
-          {/* <TabPane tab="Appellate Cases" key="2">
-            <div>
-              Filters:{' '}
-              {processFilters(appFilters).map(filter => (
-                // ! My code starting below
-                <Tag key={filter.value}>
-                  {filter.value}{' '}
-                  <span
-                    style={{ cursor: 'pointer' }}
-                    onClick={() => removeSearchTerm(initialFilters, filter.key)}
-                  >
-                    X
-                  </span>
-                  //! end of my code
-                </Tag>
-              ))}
-            </div>
-            <div className="case-table">
-              <Table
-                className="cases_table appCases"
-                rowSelection={rowSelection}
-                rowKey={record => record.case_id}
-                columns={columns}
-                dataSource={appCases}
-                onChange={changeSorter}
-              />
-            </div>
-          </TabPane> */}
           <TabPane
             tab={
               <Button
