+++ src/utils/format_date_util.js
@@ -1,0 +1,6 @@
+export function formatDate(text) {
+  var year = text.slice(0, 4);
+  var month = text.slice(5, 7);
+  var day = text.slice(8, 10);
+  return month + '/' + day + '/' + year;
+}
